+++ client/src/hooks/index.tsx
@@ -1,2 +1,3 @@
 export { useLogin } from "./useLogin";
 export { useSignup } from "./useSignup";
+export { useMenuSideabar } from "./useMenuSidebar";
